Restructure test arm code
