+++ src/main/java/com/stuypulse/robot/test/TestbotContainer.java
@@ -4,6 +4,7 @@ import com.stuypulse.robot.constants.Ports;
 import com.stuypulse.robot.subsystems.Pump;
 import com.stuypulse.robot.util.BootlegXbox;
 import com.stuypulse.stuylib.input.Gamepad;
+import com.stuypulse.stuylib.network.SmartBoolean;
 import com.stuypulse.stuylib.network.SmartNumber;
 
 import edu.wpi.first.cameraserver.CameraServer;
@@ -21,9 +22,7 @@ public class TestbotContainer {
     // // Subsystem
     public final TestIntake intake = new TestIntake();
     public final TestSwerveDrive swerve = new TestSwerveDrive();
-    // public final TestArm arm = new TestArm();
-    // public final TestArmPD arm = new TestArmPD();
-    public final TestArmDyn arm = new TestArmDyn();
+    public final TestArm arm = new TestArm();
     public final TestPlant plant = new TestPlant();
     public final TestWings wings = new TestWings();
     
@@ -100,23 +99,27 @@ public class TestbotContainer {
 
     private final SmartNumber SHOULDER_VOLTS = new SmartNumber("Arm/Shoulder Input Volts", 0);
     private final SmartNumber WRIST_VOLTS = new SmartNumber("Arm/Wrist Input Volts", 0);
+    private final SmartBoolean ARM_DRIVE = new SmartBoolean("Arm/Arm Drive", false);
 
     private void configureDefaultCommands() {
         // arm
-        // arm.setDefaultCommand(arm.run(() -> {
-        //     arm.runShoulder(MathUtil.clamp(SHOULDER_VOLTS.get(), -3, 3));
-        //     arm.runWrist(MathUtil.clamp(WRIST_VOLTS.get(), -5, 5));
-        // }));
-        // arm.setDefaultCommand(arm.run(() -> {
-        //     double shoulderVolts = MathUtil.applyDeadband(operator.getLeftY(), 0.05) * 3;
-        //     double wristVolts = MathUtil.applyDeadband(operator.getRightY(), 0.05) * 5;
-
-        //     SmartDashboard.putNumber("Arm/Shoulder Voltage", shoulderVolts);
-        //     SmartDashboard.putNumber("Arm/Wrist Voltage", wristVolts);
-
-        //     arm.runShoulder(shoulderVolts);
-        //     arm.runWrist(wristVolts);
-        // }));
+        arm.setDefaultCommand(arm.run(() -> {
+            if (!SmartDashboard.getBoolean("Arm/Setpoint Control", false)) {
+                if (ARM_DRIVE.get()) {
+                    double shoulderVolts = MathUtil.applyDeadband(operator.getLeftY(), 0.05) * 3;
+                    double wristVolts = MathUtil.applyDeadband(operator.getRightY(), 0.05) * 5;
+        
+                    SmartDashboard.putNumber("Arm/Shoulder Voltage", shoulderVolts);
+                    SmartDashboard.putNumber("Arm/Wrist Voltage", wristVolts);
+        
+                    arm.runShoulder(shoulderVolts);
+                    arm.runWrist(wristVolts);
+                } else {
+                    arm.runShoulder(MathUtil.clamp(SHOULDER_VOLTS.get(), -3, 3));
+                    arm.runWrist(MathUtil.clamp(WRIST_VOLTS.get(), -5, 5));
+                }
+            }
+        }));
     }
     
 }
